DSRC l2 handler and passing in l2 type to xnetjd
Xnetjd now takes a --layer2 option to specify the l2 type on the
access point. The client just accepts the l2 specified by the access
point beacon so it doesn't need to be told.

A new DSRC handler with a separate set of retransmission rate and
number of attempts was created. It is pretty similar to the Ethernet
handler but can diverge in future.